% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenePos.fun.R
\name{GenePos}
\alias{GenePos}
\title{GenePos function}
\usage{
GenePos(x, lines_HL, isolines_HL, isolines_IT, lines_IT, line.name = FALSE)
}
\arguments{
\item{x}{REQUIRED is sequential order of the line to test, or the name with "" around it if line.name=TRUE}

\item{lines_HL}{REQUIRED is the data frame for lines converted to H and L}

\item{isolines_HL}{REQUIRED is the data frame for isolines converted to H and L}

\item{isolines_IT}{REQUIRED is the data frame for lines with raw IT values}

\item{lines_IT}{REQUIRED is the data frame for isolines with raw IT values}

\item{line.name}{OPTIONAL Defulat is FALSE, if set to TRUE then can specify the name of the line to select}
}
\description{
returns the postulated genes and the name of the line Function imports the package "matrixStats"
}
