% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SrIT.Convert.fun.R
\name{SrIT.Convert}
\alias{SrIT.Convert}
\title{SrIT.Convert function}
\usage{
SrIT.Convert(x, thresh)
}
\arguments{
\item{x}{REQUIRED x is the name of thex is a matrix containing the raw IT values to convert to H or L calls,
must be read in as a matrix as given by the example files
must be a matrix with the race name set as the row names
output is a matrix with race names set as row names}

\item{thresh}{REQUIRED thresh is the threshold IT for calling an line as high or low lines > or equal to thresh will be called high
thresh can be set to one of the following to any of the Bowden IT double values ie "44" "3-2" ";;" "0;0;"}
}
\description{
This function converts stem rust 0-4 infection types to a numeric scale
 SrIT.Convert(x, thresh)
Function SrIT.Convert(x, thresh) converts raw stem rust IT values first to weighted average
data is then converted to H and L values based on the value of thresh
data must conform to the two IT value system as described by Bowden
can be used for both stem rust and leaf rust using the 0-4 stakman scale
output is a matrix with race names set as row names
}
