% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YrIT.Convert.fun.R
\name{YrIT.Convert}
\alias{YrIT.Convert}
\title{YrIT.Convert function}
\usage{
YrIT.Convert(x, thresh)
}
\arguments{
\item{x}{REQUIRED x is the name of thex is a matrix containing the raw IT values to convert to H or L calls,
must be read in as a matrix as given by the example files
must be a matrix with the race name set as the row names
output is a matrix with race names set as row names}

\item{thresh}{REQUIRED thresh is the threshold IT for calling an line as high or low lines > or equal to thresh will be called high
thresh can be set to one any nummber between 0 and 9}
}
\description{
This function converts yellow rust 0-9 infection types to a numeric scale
converts raw stripe rust IT values to H and L values
output is a matrix with race names set as row names
}
